+++ src/v5/sink.rs
@@ -128,7 +128,7 @@ impl<'a> PublishBuilder<'a> {
     }
 
     /// Send publish packet with QoS 0
-    pub fn at_most_once(&mut self) {
+    pub fn send_at_most_once(&mut self) {
         if let Some(packet) = self.packet.take() {
             if let Some(ref sink) = self.sink.0.borrow().sink {
                 log::trace!("Publish (QoS-0) to {:?}", packet.topic);
@@ -142,7 +142,7 @@ impl<'a> PublishBuilder<'a> {
     }
 
     /// Send publish packet with QoS 1
-    pub fn at_least_once(&mut self) -> impl Future<Output = Result<(), ()>> {
+    pub fn send_at_least_once(&mut self) -> impl Future<Output = Result<(), ()>> {
         if let Some(mut packet) = self.packet.take() {
             let mut inner = self.sink.0.borrow_mut();